pyproject.toml
[tool.poetry]
name = "chainmapper"
version = "0.1.0"
description = ""
authors = ["jinkai-b <jinkai-b@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
chainmapper = 'chainmapper.convert:main'

[tool.poetry.dependencies]
python = "^3.12"
python-evtx = "^0.7.4"
lxml = "^5.3.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
